Remove derive_builder, replace by Job builder itself logic
