+++ aj/examples/cron_job.rs
@@ -4,7 +4,7 @@ use aj::{
         actix_rt::time::sleep,
         serde::{Deserialize, Serialize},
     },
-    main, BackgroundJob, Executable, JobBuilder, JobContext, AJ,
+    main, BackgroundJob, Executable, JobContext, AJ,
 };
 use aj_core::get_now_as_secs;
 
@@ -26,10 +26,8 @@ async fn main() {
 
     println!("Start time {}", get_now_as_secs());
     let _ = AJob
-        .job_builder()
+        .job()
         .cron("* * * * * * *")
-        .build()
-        .unwrap()
         .run()
         .await;
