+++ aj_core/Cargo.toml
@@ -31,5 +31,4 @@ log = "0.4"
 lazy_static = { version = "1.4.0" }
 async-trait = "0.1.74"
 dashmap = "5.5.3"
-derive_builder = "0.20.0"
 tokio = "1.23.1"
