+++ aj_core/src/lib.rs
@@ -4,8 +4,6 @@
 extern crate log;
 #[macro_use]
 extern crate serde_with;
-#[macro_use]
-extern crate derive_builder;
 
 pub mod aj;
 pub mod backend;
